refactor: pagination component as separate client component
