+++ app/tag-data.json
@@ -1,1 +1,1 @@
-{"redux":1,"react":5,"architecture":1,"frontend":4,"vue":1,"angular":3,"statemanagement":1,"codinghorror":1,"forms":1,"accessiblity":1,"html5":1,"webstandards":1,"webcomponents":2,"senior":4,"career":12,"productivity":9,"leadership":1,"responsivedesign":1,"testing":6,"agile":2,"teamwork":1,"management":1,"prioritization":1,"jest":1,"typescript":2,"staff":1,"wishiknew":2,"next-js":2,"tailwind":1,"guide":2,"feature":1,"beginner":11,"bugs":1,"engineeringnotes":1,"mentalhealth":1,"impostersyndrome":1,"growth":1,"webperf":2,"javascript":4,"intermediate":3,"git":2,"showdev":7,"discuss":1,"shecoded":4,"storybook":1,"beginners":1,"webdev":4,"virtualmachine":1,"ie11":1,"devops":1,"aws":1,"shecodedally":1,"css":1,"deploy":1,"nearlyfreespeech":1,"debugging":1,"sql":1,"node":1,"heroku":1,"mongodb":1,"backend":1,"layoff":1,"interview":1,"email":1,"outlook":1}
+{"beginner":11,"career":12,"redux":1,"react":5,"architecture":1,"frontend":3,"vue":1,"angular":3,"statemanagement":1,"codinghorror":1,"testing":6,"beginners":1,"webdev":4,"javascript":4,"productivity":9,"discuss":1,"bugs":1,"agile":2,"engineeringnotes":1,"webcomponents":1,"showdev":7,"devops":1,"aws":1,"senior":4,"staff":1,"wishiknew":2,"typescript":2,"webperf":2,"jest":1,"teamwork":1,"management":1,"prioritization":1,"email":1,"outlook":1,"git":2,"next-js":1,"guide":1,"deploy":1,"nearlyfreespeech":1,"shecoded":4,"css":1,"interview":1,"layoff":1,"mentalhealth":1,"impostersyndrome":1,"growth":1,"mongodb":1,"backend":1,"intermediate":3,"virtualmachine":1,"ie11":1,"responsivedesign":1,"storybook":1,"leadership":1,"node":1,"heroku":1,"debugging":1,"sql":1,"shecodedally":1}
