+++ components/Pagination.tsx
@@ -1,0 +1,60 @@
+'use client'
+
+import { usePathname } from "next/navigation"
+import React from "react"
+import Link from "./Link"
+export interface PaginationProps {
+    totalPages: number
+    currentPage: number
+}
+
+export function Pagination({ totalPages, currentPage }: PaginationProps) {
+    const pathname = usePathname()
+    const basePath = pathname.split('/')[1]
+    const prevPage = currentPage - 1 > 0
+    const nextPage = currentPage + 1 <= totalPages
+
+    return (
+        <div className="space-y-2 pb-8 pt-6 md:space-y-5">
+            <nav className="flex justify-between">
+                {!prevPage && (
+                    <button
+                        className="cursor-auto disabled:opacity-50"
+                        disabled={!prevPage}
+                    >
+                        Previous
+                    </button>
+                )}
+                {prevPage && (
+                    <Link
+                        href={
+                            currentPage - 1 === 1
+                                ? `/${basePath}/`
+                                : `/${basePath}/page/${currentPage - 1}`
+                        }
+                        rel="prev"
+                    >
+                        &larr; Previous
+                    </Link>
+                )}
+                <span>
+                    {currentPage} of {totalPages}
+                </span>
+                {!nextPage && (
+                    <button
+                        className="cursor-auto disabled:opacity-50"
+                        disabled={!nextPage}
+                    >
+                        Next
+                    </button>
+                )}
+                {nextPage && (
+                    <Link href={`/${basePath}/page/${currentPage + 1}`} rel="next">
+                        Next &rarr;
+                    </Link>
+                )}
+            </nav>
+        </div>
+    )
+}
+export default Pagination
