+++ scripts/rss.mjs
@@ -15,7 +15,7 @@ const generateRssItem = (config, post) => `
     <title>${escape(post.title)}</title>
     <link>${config.siteUrl}/blog/${post.slug}</link>
     ${post.summary && `<description>${escape(post.summary)}</description>`}
-    <pubDate>${new Date(post?.date).toUTCString()}</pubDate>
+    <pubDate>${new Date(post?.date)?.toUTCString()}</pubDate>
     <author>${config.email}${config.author}</author>
     ${post.tags && post.tags.map((t) => `<category>${t}</category>`).join('')}
   </item>
@@ -30,7 +30,7 @@ const generateRss = (config, posts, page = 'feed.xml') => `
       <language>${config.language}</language>
       <managingEditor>${config.email} (${config.author})</managingEditor>
       <webMaster>${config.email} (${config.author})</webMaster>
-      <lastBuildDate>${new Date(posts[0]?.date).toUTCString()}</lastBuildDate>
+      <lastBuildDate>${new Date(posts[0]?.date)?.toUTCString()}</lastBuildDate>
       <atom:link href="${config.siteUrl}/${page}" rel="self" type="application/rss+xml"/>
       ${posts.map((post) => generateRssItem(config, post)).join('')}
     </channel>
